+++ test/smokes/golangci_lint/expectations.rb
@@ -1,6 +1,6 @@
 s = Runners::Testing::Smoke
 
-default_version = "1.41.1"
+default_version = "1.42.0"
 
 s.add_test(
   "target",