+++ test/smokes/rubocop/expectations.rb
@@ -1,6 +1,6 @@
 s = Runners::Testing::Smoke
 
-default_version = "1.19.1"
+default_version = "1.20.0"
 
 s.add_test(
   "sandbox_rails",