+++ test/smokes/slim_lint/expectations.rb
@@ -1,6 +1,6 @@
 s = Runners::Testing::Smoke
 
-default_version = "0.22.0"
+default_version = "0.22.1"
 
 s.add_test(
   "success",