+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 Updated tools:
 
 - **GolangCI-Lint** 1.41.1 -> 1.42.0 [#2619](https://github.com/sider/runners/pull/2619)
+- **Pylint** 2.9.6 -> 2.10.2 [#2621](https://github.com/sider/runners/pull/2621)
 
 Misc: