+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Updated tools:
 - **GolangCI-Lint** 1.41.1 -> 1.42.0 [#2619](https://github.com/sider/runners/pull/2619)
 - **hadolint** 2.6.0 -> 2.7.0 [#2639](https://github.com/sider/runners/pull/2639)
 - **Pylint** 2.9.6 -> 2.10.2 [#2621](https://github.com/sider/runners/pull/2621)
+- **RuboCop** 1.19.1 -> 1.20.0 [#2640](https://github.com/sider/runners/pull/2640)
 
 Misc: