+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Updated environments:
 Updated tools:
 
 - **CoffeeLint** 5.0.1 -> 5.0.2 [#2631](https://github.com/sider/runners/pull/2631)
+- **detekt** 1.18.0 -> 1.18.1 [#2662](https://github.com/sider/runners/pull/2662)
 - **GolangCI-Lint** 1.41.1 -> 1.42.0 [#2619](https://github.com/sider/runners/pull/2619)
 - **hadolint** 2.6.0 -> 2.7.0 [#2639](https://github.com/sider/runners/pull/2639)
 - **PMD CPD** 6.37.0 -> 6.38.0 [#2647](https://github.com/sider/runners/pull/2647)